[UDT-229] refactor: contentform을 5개 파일로 분리 (basicinfo, detailedinfo, castinfo, directorinfo, platforminfo)
